validation

+++ server/src/services/categoryService.js
@@ -10,6 +10,10 @@ const createCategory = async (data) => {
   });
 };
 const getCategory = async (id) => {
+  const category =await validateCategoryById(id);
+  if (!category) {
+    throw new Error("Category not found");
+  }
   return prisma.category.findFirstOrThrow({
     where : {
         'id' : parseInt(id)
@@ -18,6 +22,10 @@ const getCategory = async (id) => {
 };
 
 const updateCategory = async (id ,data) => {
+  const category =await validateCategoryById(id);
+  if (!category) {
+    throw new Error("Category not found");
+  }
   return prisma.category.update(
     {
       where: { id: parseInt(id) },
@@ -27,12 +35,32 @@ const updateCategory = async (id ,data) => {
 };
 
 const deleteCategory = async (id ,data) => {
+ const category =await validateCategoryById(id);
+  if (!category) {
+    throw new Error("Category not found");
+  }
+  
   return prisma.category.delete(
     {
       where: { id: parseInt(id) },
     }
   )
 };
+
+const validateCategoryById = async (id) => {
+  await prisma.product.deleteMany({
+    where: {
+      categoryId: Number(id)
+    }
+  });
+  const category = await prisma.category.findUnique({
+    where: {
+      id: Number(id)
+    }
+  });
+  return category;
+};
+
 
 module.exports = {
   getCategories,
